Solution for 2015, Day 23

+++ common-utils/src/main/java/me/zodac/advent/util/MathUtils.java
@@ -69,23 +69,23 @@ public final class MathUtils {
     }
 
     /**
-     * Checks if the input {@code int} is even.
+     * Checks if the input {@code long} is even.
      *
-     * @param value the {@code int} to check
-     * @return {@code true} if the {@code int} is even
+     * @param value the {@code long} to check
+     * @return {@code true} if the {@code long} is even
      */
-    public static boolean isEven(final int value) {
+    public static boolean isEven(final long value) {
         return value % 2 == 0;
     }
 
     /**
-     * Checks if the input {@code int} is odd.
+     * Checks if the input {@code long} is odd.
      *
-     * @param value the {@code int} to check
-     * @return {@code true} if the {@code int} is odd
-     * @see #isEven(int)
+     * @param value the {@code long} to check
+     * @return {@code true} if the {@code long} is odd
+     * @see #isEven(long)
      */
-    public static boolean isOdd(final int value) {
+    public static boolean isOdd(final long value) {
         return !isEven(value);
     }
